% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Monte Carlo simulation function
Predicting Theoretical Value of Options per Share Using Monte Carlo Simulations}
\usage{
monte_carlo(I, M, S_0, K, Time, r, sigma)
}
\arguments{
\item{I}{number ofsimulation}

\item{M}{number of time steps}

\item{S_0}{The initial price of the underlying stock}

\item{K}{Exercise price (conversion price)}

\item{Time}{Simulate paths over time intervals}

\item{r}{risk free rate}

\item{sigma}{Volatility (Standard Deviation of Return)}
}
\value{
No return value, called for side effects
}
\description{
Monte Carlo simulation function
Predicting Theoretical Value of Options per Share Using Monte Carlo Simulations
}
\examples{
monte_carlo(I=10000,M=50,S_0=5.9,K=5.43,T=1.353,r=0.018482,sigma=0.2616)
}
