% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_value.R
\name{option_value}
\alias{option_value}
\title{Option_value functuon
Calculate the four value comparisons:Option value of convertible bond,Theoretical value of convertible bonds (pure bond value + option value),The difference between the theoretical price of convertible bonds and the current price,The ratio of the difference between the theoretical price of convertible bonds and the current price}
\usage{
option_value(value_per, current_price, conver_price, netdebt_value)
}
\arguments{
\item{value_per}{Option value per share(numeric)}

\item{current_price}{Current price of convertible bonds}

\item{conver_price}{Conversion price}

\item{netdebt_value}{Pure debt value}
}
\value{
No return value, called for side effects
}
\description{
Option_value functuon
Calculate the four value comparisons:Option value of convertible bond,Theoretical value of convertible bonds (pure bond value + option value),The difference between the theoretical price of convertible bonds and the current price,The ratio of the difference between the theoretical price of convertible bonds and the current price
}
\examples{
option_value( value_per=1.02,current_price=122.82,conver_price=5.43,netdebt_value=104.05  )
}
